feat(#106): Split FrequencyRoomScreen into maintainable widgets
Refactor the 2,002-line FrequencyRoomScreen by extracting inline subwidgets
into separate, reusable components. Reduces main file to 999 lines (50% reduction).

**Extracted widgets:**
- now_playing_card.dart (246 lines) — media player UI with controls
- peer_drawer.dart (202 lines) — per-peer volume/mute/remove sheet
- queue_sheet.dart (154 lines) — shared media queue display
- invite_sheet.dart (126 lines) — frequency invite with QR code
- output_sheet.dart (182 lines) — audio routing selection
- push_to_talk_button.dart (51 lines) — PTT interaction widget
- peer_row.dart (100 lines) — roster list item widget

**Benefits:**
- All extracted files under 250 lines
- Self-contained widgets with clear prop interfaces
- Improved testability and reusability
- Pure refactoring with no behavior changes

Addresses issue #106 acceptance criteria: no file in lib/ exceeds 600 lines.
