+++ lib/screens/frequency_room_screen.dart
@@ -13,28 +13,15 @@ import '../services/audio_service.dart';
 import '../theme/app_theme.dart';
 import '../widgets/frequency_atoms.dart';
 import '../widgets/frequency_toast_host.dart';
+import '../widgets/invite_sheet.dart';
+import '../widgets/now_playing_card.dart';
+import '../widgets/output_sheet.dart';
+import '../widgets/peer_drawer.dart';
+import '../widgets/peer_row.dart';
+import '../widgets/push_to_talk_button.dart';
+import '../widgets/queue_sheet.dart';
 
-enum AudioOutput { bluetooth, earpiece, speaker }
-
-extension on AudioOutput {
-  String get label => switch (this) {
-        AudioOutput.bluetooth => 'Bluetooth headphones',
-        AudioOutput.earpiece => 'Phone earpiece',
-        AudioOutput.speaker => 'Phone speaker',
-      };
-
-  IconData get icon => switch (this) {
-        AudioOutput.bluetooth => Icons.bluetooth,
-        AudioOutput.earpiece => Icons.mic_none,
-        AudioOutput.speaker => Icons.volume_up,
-      };
-
-  String subFor(String bt) => switch (this) {
-        AudioOutput.bluetooth => bt.isEmpty ? 'Paired headphones' : bt,
-        AudioOutput.earpiece => 'Private, held to ear',
-        AudioOutput.speaker => 'Loud · everyone nearby hears',
-      };
-}
+export '../widgets/output_sheet.dart' show AudioOutput;
 
 /// Main "On air" room — voice + now playing.
 class FrequencyRoomScreen extends StatefulWidget {
@@ -528,10 +515,6 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
     );
   }
 
-  String _outputName() {
-    return _output == AudioOutput.bluetooth ? 'headphones' : _output.label;
-  }
-
   /// Maps a protocol-layer peer to the UI's presentation model.
   /// Hue is deterministic from peerId so the same peer always gets
   /// the same color across sessions.
@@ -604,13 +587,15 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
                 child: ListView(
                   padding: const EdgeInsets.fromLTRB(20, 16, 20, 28),
                   children: [
-                    _NowPlayingCard(
+                    NowPlayingCard(
                       track: _track,
                       source: source,
                       isPodcast: _source == 'Podcasts',
                       playing: _playing,
                       progress: _progress,
-                      lastAction: _lastAction,
+                      lastActionBy: _lastAction.by,
+                      lastActionWhat: _lastAction.action,
+                      lastActionWhen: _lastAction.when,
                       onPlay: _togglePlay,
                       onNext: _next,
                       onPrev: _prev,
@@ -676,7 +661,7 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
                                 child: Column(
                                   children: [
                                     for (int i = 0; i < peers.length; i++)
-                                      _PeerRow(
+                                      PeerRow(
                                         key: ValueKey(peers[i].id),
                                         person: peers[i],
                                         first: i == 0,
@@ -866,7 +851,7 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
             ),
           ),
           if (widget.pttMode)
-            _PushToTalkButton(
+            PushToTalkButton(
               holding: _holdingPtt,
               onChange: _setPttHolding,
             )
@@ -906,7 +891,7 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
         borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
       ),
       builder: (ctx) {
-        return _PeerDrawer(
+        return PeerDrawer(
           person: person,
           isHost: widget.isHost,
           initialVolume: _volumeFor(person.id),
@@ -945,7 +930,7 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
       shape: const RoundedRectangleBorder(
         borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
       ),
-      builder: (ctx) => _QueueSheet(
+      builder: (ctx) => QueueSheet(
         lib: _lib,
         currentIdx: _trackIdx,
         onPlay: (i) {
@@ -965,7 +950,7 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
       shape: const RoundedRectangleBorder(
         borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
       ),
-      builder: (_) => _InviteSheet(freq: widget.freq),
+      builder: (_) => InviteSheet(freq: widget.freq),
     );
   }
 
@@ -978,7 +963,7 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
       shape: const RoundedRectangleBorder(
         borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
       ),
-      builder: (_) => _OutputSheet(current: _output, btName: _me.btDevice),
+      builder: (_) => OutputSheet(current: _output, btName: _me.btDevice),
     );
     if (picked != null && mounted) {
       // Apply the audio routing change at the native layer
@@ -996,6 +981,10 @@ class _FrequencyRoomScreenState extends State<FrequencyRoomScreen> {
       }
     }
   }
+
+  String _outputName() {
+    return _output == AudioOutput.bluetooth ? 'headphones' : _output.label;
+  }
 }
 
 class _LastAction {
@@ -1005,996 +994,6 @@ class _LastAction {
   const _LastAction({required this.by, required this.action, required this.when});
 }
 
-// ── Now playing card ────────────────────────────────────────
 
-class _NowPlayingCard extends StatelessWidget {
-  final Track track;
-  final String source;
-  final bool isPodcast;
-  final bool playing;
-  final int progress;
-  final _LastAction lastAction;
-  final VoidCallback onPlay;
-  final VoidCallback onNext;
-  final VoidCallback onPrev;
-  final ValueChanged<double> onScrub;
-  final VoidCallback onOpenQueue;
 
-  const _NowPlayingCard({
-    required this.track,
-    required this.source,
-    required this.isPodcast,
-    required this.playing,
-    required this.progress,
-    required this.lastAction,
-    required this.onPlay,
-    required this.onNext,
-    required this.onPrev,
-    required this.onScrub,
-    required this.onOpenQueue,
-  });
 
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return FreqCard(
-      padding: const EdgeInsets.all(16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Expanded(
-                child: Text(
-                  'LISTENING TOGETHER · ${source.toUpperCase()}',
-                  style: TextStyle(
-                    fontFamily: 'Inter',
-                    fontSize: 10,
-                    fontWeight: FontWeight.w500,
-                    letterSpacing: 1.0,
-                    color: c.ink3,
-                  ),
-                ),
-              ),
-              Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  VuMeter(active: playing),
-                  const SizedBox(width: 5),
-                  Text(
-                    playing ? 'Live' : 'Paused',
-                    style: TextStyle(
-                      fontFamily: 'Inter',
-                      fontSize: 10,
-                      fontWeight: FontWeight.w500,
-                      letterSpacing: 1.0,
-                      color: c.ink3,
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-          const SizedBox(height: 12),
-          Row(
-            children: [
-              _Artwork(isPodcast: isPodcast),
-              const SizedBox(width: 14),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      track.tag,
-                      style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3),
-                    ),
-                    const SizedBox(height: 2),
-                    Text(
-                      track.title,
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                      style: TextStyle(
-                        fontFamily: 'Inter',
-                        fontSize: 16,
-                        fontWeight: FontWeight.w600,
-                        letterSpacing: -0.16,
-                        color: c.ink,
-                      ),
-                    ),
-                    Text(
-                      track.artist,
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                      style: TextStyle(
-                        fontFamily: 'Inter',
-                        fontSize: 13,
-                        color: c.ink2,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 14),
-          Slider(
-            value: progress.toDouble().clamp(0.0, track.durationSeconds.toDouble()).toDouble(),
-            min: 0,
-            max: track.durationSeconds.toDouble(),
-            onChanged: onScrub,
-          ),
-          Row(
-            children: [
-              Text(formatTime(progress), style: kMonoStyle.copyWith(fontSize: 11, color: c.ink3)),
-              const Spacer(),
-              Text('-${formatTime(track.durationSeconds - progress)}',
-                  style: kMonoStyle.copyWith(fontSize: 11, color: c.ink3)),
-            ],
-          ),
-          const SizedBox(height: 8),
-          Row(
-            children: [
-              GhostButton(icon: Icons.skip_previous, onPressed: onPrev),
-              const SizedBox(width: 4),
-              _PlayCircle(playing: playing, onTap: onPlay),
-              const SizedBox(width: 4),
-              GhostButton(icon: Icons.skip_next, onPressed: onNext),
-              const Spacer(),
-              FreqButton(
-                icon: Icons.queue_music,
-                label: 'Queue',
-                padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                fontSize: 13,
-                onPressed: onOpenQueue,
-              ),
-            ],
-          ),
-          const SizedBox(height: 12),
-          Container(
-            padding: const EdgeInsets.only(top: 10),
-            decoration: BoxDecoration(
-              border: Border(top: BorderSide(color: c.line)),
-            ),
-            child: Row(
-              children: [
-                Expanded(
-                  child: RichText(
-                    text: TextSpan(
-                      style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3),
-                      children: [
-                        TextSpan(
-                          text: lastAction.by,
-                          style: TextStyle(color: c.ink2, fontWeight: FontWeight.w500),
-                        ),
-                        TextSpan(text: ' ${lastAction.action}'),
-                      ],
-                    ),
-                  ),
-                ),
-                Text(
-                  lastAction.when,
-                  style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _Artwork extends StatelessWidget {
-  final bool isPodcast;
-  const _Artwork({required this.isPodcast});
-
-  @override
-  Widget build(BuildContext context) {
-    final base = isPodcast
-        ? const [Color(0xFFEFD8C3), Color(0xFFF6E6D6)]
-        : const [Color(0xFFD4D5EE), Color(0xFFE2E3F4)];
-    return Container(
-      width: 64,
-      height: 64,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(10),
-        gradient: LinearGradient(
-          colors: base,
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          tileMode: TileMode.repeated,
-        ),
-      ),
-      alignment: Alignment.center,
-      child: Icon(
-        isPodcast ? Icons.podcasts : Icons.music_note,
-        size: 22,
-        color: isPodcast ? const Color(0xFF704D2A) : const Color(0xFF3F3F73),
-      ),
-    );
-  }
-}
-
-class _PlayCircle extends StatelessWidget {
-  final bool playing;
-  final VoidCallback onTap;
-  const _PlayCircle({required this.playing, required this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Material(
-      color: c.ink,
-      shape: const CircleBorder(),
-      child: InkWell(
-        customBorder: const CircleBorder(),
-        onTap: onTap,
-        child: SizedBox(
-          width: 48,
-          height: 48,
-          child: Icon(
-            playing ? Icons.pause : Icons.play_arrow,
-            size: 22,
-            color: c.bg,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// ── Push-to-talk ────────────────────────────────────────────
-
-class _PushToTalkButton extends StatelessWidget {
-  final bool holding;
-  final ValueChanged<bool> onChange;
-  const _PushToTalkButton({required this.holding, required this.onChange});
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Listener(
-      onPointerDown: (_) => onChange(true),
-      onPointerUp: (_) => onChange(false),
-      onPointerCancel: (_) => onChange(false),
-      child: AnimatedContainer(
-        duration: const Duration(milliseconds: 120),
-        padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 8),
-        decoration: BoxDecoration(
-          color: holding ? c.accent : c.surface,
-          borderRadius: BorderRadius.circular(10),
-          border: Border.all(color: holding ? Colors.transparent : c.line),
-        ),
-        constraints: const BoxConstraints(minWidth: 104),
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(Icons.mic, size: 14, color: holding ? c.accentInk : c.ink),
-            const SizedBox(width: 6),
-            Text(
-              holding ? 'On air' : 'Hold to talk',
-              style: TextStyle(
-                fontFamily: 'Inter',
-                fontSize: 13,
-                fontWeight: FontWeight.w600,
-                color: holding ? c.accentInk : c.ink,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// ── Peer row ────────────────────────────────────────────────
-
-class _PeerRow extends StatelessWidget {
-  final Person person;
-  final bool first;
-  final bool talking;
-  final bool muted;
-  final double volume;
-  final VoidCallback onTap;
-
-  const _PeerRow({
-    super.key,
-    required this.person,
-    required this.first,
-    required this.talking,
-    required this.muted,
-    required this.volume,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Material(
-      color: c.surface,
-      child: InkWell(
-        onTap: onTap,
-        child: Container(
-          padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 12),
-          decoration: BoxDecoration(
-            border: Border(
-              top: first ? BorderSide.none : BorderSide(color: c.line),
-            ),
-          ),
-          child: Row(
-            children: [
-              FreqAvatar(person: person, size: 36, talking: talking, muted: muted),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      person.name,
-                      style: TextStyle(
-                        fontFamily: 'Inter',
-                        fontSize: 14,
-                        fontWeight: FontWeight.w500,
-                        color: c.ink,
-                      ),
-                    ),
-                    Row(
-                      children: [
-                        Icon(Icons.bluetooth, size: 10, color: c.ink3),
-                        const SizedBox(width: 4),
-                        Flexible(
-                          child: Text(
-                            person.btDevice,
-                            overflow: TextOverflow.ellipsis,
-                            style: TextStyle(
-                              fontFamily: 'Inter',
-                              fontSize: 12,
-                              color: c.ink3,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-              if (talking)
-                Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    VuMeter(color: c.accent),
-                    const SizedBox(width: 5),
-                    Text('talking', style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.accent)),
-                  ],
-                )
-              else if (muted)
-                Text('muted', style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3))
-              else
-                Text(
-                  '${(volume * 100).round()}%',
-                  style: kMonoStyle.copyWith(fontSize: 11, color: c.ink3),
-                ),
-              const SizedBox(width: 6),
-              Icon(Icons.chevron_right, size: 16, color: c.ink3),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// ── Peer drawer ────────────────────────────────────────────
-
-class _PeerDrawer extends StatefulWidget {
-  final Person person;
-  final bool isHost;
-  final double initialVolume;
-  final bool initialMuted;
-  final void Function(double volume, bool muted) onChanged;
-  final VoidCallback onRemove;
-
-  const _PeerDrawer({
-    required this.person,
-    required this.isHost,
-    required this.initialVolume,
-    required this.initialMuted,
-    required this.onChanged,
-    required this.onRemove,
-  });
-
-  @override
-  State<_PeerDrawer> createState() => _PeerDrawerState();
-}
-
-class _PeerDrawerState extends State<_PeerDrawer> {
-  late double _volume = widget.initialVolume;
-  late bool _muted = widget.initialMuted;
-
-  void _emit() => widget.onChanged(_volume, _muted);
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Padding(
-      padding: EdgeInsets.fromLTRB(
-        20,
-        8,
-        20,
-        MediaQuery.of(context).viewInsets.bottom + 28,
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Center(
-            child: Container(
-              width: 36,
-              height: 4,
-              margin: const EdgeInsets.only(top: 8, bottom: 14),
-              decoration: BoxDecoration(
-                color: c.line2,
-                borderRadius: BorderRadius.circular(999),
-              ),
-            ),
-          ),
-          Row(
-            children: [
-              FreqAvatar(person: widget.person, size: 44),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      widget.person.name,
-                      style: TextStyle(
-                        fontFamily: 'Inter',
-                        fontSize: 16,
-                        fontWeight: FontWeight.w600,
-                        color: c.ink,
-                      ),
-                    ),
-                    Row(
-                      children: [
-                        Icon(Icons.bluetooth, size: 11, color: c.ink3),
-                        const SizedBox(width: 4),
-                        Text(
-                          widget.person.btDevice,
-                          style: TextStyle(fontFamily: 'Inter', fontSize: 12, color: c.ink3),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-              GhostButton(icon: Icons.close, onPressed: () => Navigator.pop(context)),
-            ],
-          ),
-          const SizedBox(height: 18),
-          Container(
-            padding: const EdgeInsets.symmetric(vertical: 12),
-            decoration: BoxDecoration(border: Border(top: BorderSide(color: c.line))),
-            child: Row(
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        'Mute from your side',
-                        style: TextStyle(
-                          fontFamily: 'Inter',
-                          fontSize: 14,
-                          fontWeight: FontWeight.w500,
-                          color: c.ink,
-                        ),
-                      ),
-                      Text(
-                        'Only you stop hearing them',
-                        style: TextStyle(fontFamily: 'Inter', fontSize: 12, color: c.ink3),
-                      ),
-                    ],
-                  ),
-                ),
-                FreqSwitch(
-                  value: _muted,
-                  onChanged: (v) {
-                    setState(() => _muted = v);
-                    _emit();
-                  },
-                ),
-              ],
-            ),
-          ),
-          Container(
-            padding: const EdgeInsets.fromLTRB(0, 14, 0, 6),
-            decoration: BoxDecoration(border: Border(top: BorderSide(color: c.line))),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  crossAxisAlignment: CrossAxisAlignment.baseline,
-                  textBaseline: TextBaseline.alphabetic,
-                  children: [
-                    Expanded(
-                      child: Text(
-                        'Their voice volume',
-                        style: TextStyle(
-                          fontFamily: 'Inter',
-                          fontSize: 14,
-                          fontWeight: FontWeight.w500,
-                          color: c.ink,
-                        ),
-                      ),
-                    ),
-                    Text(
-                      '${(_volume * 100).round()}%',
-                      style: kMonoStyle.copyWith(fontSize: 12, color: c.ink3),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 10),
-                Row(
-                  children: [
-                    Icon(Icons.volume_up, size: 16, color: c.ink3),
-                    const SizedBox(width: 10),
-                    Expanded(
-                      child: Slider(
-                        value: _volume,
-                        onChanged: (v) {
-                          setState(() => _volume = v);
-                          _emit();
-                        },
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-          if (widget.isHost) ...[
-            const SizedBox(height: 14),
-            Container(
-              padding: const EdgeInsets.only(top: 14),
-              decoration: BoxDecoration(border: Border(top: BorderSide(color: c.line))),
-              child: Column(
-                children: [
-                  FreqButton(
-                    label: 'Remove from frequency',
-                    block: true,
-                    labelColor: c.danger,
-                    padding: const EdgeInsets.symmetric(vertical: 12),
-                    onPressed: widget.onRemove,
-                  ),
-                  const SizedBox(height: 8),
-                  Text(
-                    'As host, only you can remove people.',
-                    style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ],
-      ),
-    );
-  }
-}
-
-// ── Queue sheet ─────────────────────────────────────────────
-
-class _QueueSheet extends StatelessWidget {
-  final MediaSourceLib lib;
-  final int currentIdx;
-  final ValueChanged<int> onPlay;
-
-  const _QueueSheet({required this.lib, required this.currentIdx, required this.onPlay});
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return DraggableScrollableSheet(
-      initialChildSize: 0.72,
-      minChildSize: 0.4,
-      maxChildSize: 0.92,
-      expand: false,
-      builder: (_, controller) {
-        return Padding(
-          padding: const EdgeInsets.fromLTRB(20, 8, 20, 28),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Center(
-                child: Container(
-                  width: 36,
-                  height: 4,
-                  margin: const EdgeInsets.only(top: 8, bottom: 14),
-                  decoration: BoxDecoration(
-                    color: c.line2,
-                    borderRadius: BorderRadius.circular(999),
-                  ),
-                ),
-              ),
-              Row(
-                children: [
-                  Expanded(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          'Shared queue',
-                          style: TextStyle(
-                            fontFamily: 'Inter',
-                            fontSize: 16,
-                            fontWeight: FontWeight.w600,
-                            color: c.ink,
-                          ),
-                        ),
-                        Text(
-                          'From ${lib.name} · everyone can reorder',
-                          style: TextStyle(fontFamily: 'Inter', fontSize: 12, color: c.ink3),
-                        ),
-                      ],
-                    ),
-                  ),
-                  GhostButton(icon: Icons.close, onPressed: () => Navigator.pop(context)),
-                ],
-              ),
-              const SizedBox(height: 12),
-              Expanded(
-                child: ListView.builder(
-                  controller: controller,
-                  itemCount: lib.queue.length,
-                  itemBuilder: (_, i) {
-                    final t = lib.queue[i];
-                    final current = i == currentIdx;
-                    return InkWell(
-                      onTap: () => onPlay(i),
-                      child: Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 12),
-                        decoration: BoxDecoration(
-                          border: Border(
-                            top: i == 0 ? BorderSide.none : BorderSide(color: c.line),
-                          ),
-                        ),
-                        child: Row(
-                          children: [
-                            SizedBox(
-                              width: 28,
-                              child: Center(
-                                child: current
-                                    ? VuMeter(color: c.accent)
-                                    : Text(
-                                        (i + 1).toString().padLeft(2, '0'),
-                                        style: kMonoStyle.copyWith(fontSize: 12, color: c.ink3),
-                                      ),
-                              ),
-                            ),
-                            const SizedBox(width: 12),
-                            Expanded(
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                  Text(
-                                    t.title,
-                                    maxLines: 1,
-                                    overflow: TextOverflow.ellipsis,
-                                    style: TextStyle(
-                                      fontFamily: 'Inter',
-                                      fontSize: 14,
-                                      fontWeight: current ? FontWeight.w600 : FontWeight.w500,
-                                      color: c.ink,
-                                    ),
-                                  ),
-                                  Text(
-                                    t.artist,
-                                    maxLines: 1,
-                                    overflow: TextOverflow.ellipsis,
-                                    style: TextStyle(
-                                      fontFamily: 'Inter',
-                                      fontSize: 12,
-                                      color: c.ink3,
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                            const SizedBox(width: 12),
-                            Text(
-                              formatTime(t.durationSeconds),
-                              style: kMonoStyle.copyWith(fontSize: 12, color: c.ink3),
-                            ),
-                          ],
-                        ),
-                      ),
-                    );
-                  },
-                ),
-              ),
-              const SizedBox(height: 10),
-              FreqButton(
-                icon: Icons.add,
-                label: 'Add from ${lib.name}',
-                block: true,
-                padding: const EdgeInsets.symmetric(vertical: 12),
-                onPressed: () {},
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
-
-// ── Invite sheet ────────────────────────────────────────────
-
-class _InviteSheet extends StatefulWidget {
-  final String freq;
-  const _InviteSheet({required this.freq});
-
-  @override
-  State<_InviteSheet> createState() => _InviteSheetState();
-}
-
-class _InviteSheetState extends State<_InviteSheet> {
-  bool _copied = false;
-  Timer? _copiedReset;
-
-  @override
-  void dispose() {
-    _copiedReset?.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(20, 8, 20, 28),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Container(
-            width: 36,
-            height: 4,
-            margin: const EdgeInsets.only(top: 8, bottom: 14),
-            decoration: BoxDecoration(
-              color: c.line2,
-              borderRadius: BorderRadius.circular(999),
-            ),
-          ),
-          Text(
-            'INVITE NEARBY',
-            style: TextStyle(
-              fontFamily: 'Inter',
-              fontSize: 11,
-              fontWeight: FontWeight.w500,
-              letterSpacing: 1.1,
-              color: c.ink3,
-            ),
-          ),
-          const SizedBox(height: 8),
-          Text(
-            widget.freq,
-            style: kMonoStyle.copyWith(
-              fontSize: 48,
-              fontWeight: FontWeight.w500,
-              letterSpacing: -0.96,
-              color: c.ink,
-              height: 1,
-            ),
-          ),
-          const SizedBox(height: 4),
-          Text(
-            'MHz · your Frequency',
-            style: TextStyle(fontFamily: 'Inter', fontSize: 12, color: c.ink3),
-          ),
-          const SizedBox(height: 20),
-          Container(
-            width: 160,
-            height: 160,
-            decoration: BoxDecoration(
-              color: c.surface,
-              borderRadius: BorderRadius.circular(16),
-              border: Border.all(color: c.line),
-            ),
-            padding: const EdgeInsets.all(10),
-            child: GridView.builder(
-              physics: const NeverScrollableScrollPhysics(),
-              gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                crossAxisCount: 9,
-                mainAxisSpacing: 2,
-                crossAxisSpacing: 2,
-              ),
-              itemCount: 81,
-              itemBuilder: (_, i) {
-                final on = ((i * 37 + 7) % 3) == 0 ||
-                    [0, 1, 7, 8, 9, 17, 63, 64, 65, 71, 72].contains(i);
-                return Container(
-                  decoration: BoxDecoration(
-                    color: on ? c.ink : Colors.transparent,
-                    borderRadius: BorderRadius.circular(1),
-                  ),
-                );
-              },
-            ),
-          ),
-          const SizedBox(height: 12),
-          FreqButton(
-            block: true,
-            icon: _copied ? Icons.check : Icons.copy,
-            label: _copied ? 'Copied invite' : 'Copy invite link',
-            padding: const EdgeInsets.symmetric(vertical: 12),
-            onPressed: () {
-              setState(() => _copied = true);
-              _copiedReset?.cancel();
-              _copiedReset = Timer(const Duration(milliseconds: 1600), () {
-                if (mounted) setState(() => _copied = false);
-              });
-            },
-          ),
-          const SizedBox(height: 12),
-          Text(
-            'Anyone within ~30m can tune in.',
-            style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// ── Output sheet ───────────────────────────────────────────
-
-class _OutputSheet extends StatelessWidget {
-  final AudioOutput current;
-  final String btName;
-  const _OutputSheet({required this.current, required this.btName});
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(20, 8, 20, 28),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Center(
-            child: Container(
-              width: 36,
-              height: 4,
-              margin: const EdgeInsets.only(top: 8, bottom: 14),
-              decoration: BoxDecoration(
-                color: c.line2,
-                borderRadius: BorderRadius.circular(999),
-              ),
-            ),
-          ),
-          Row(
-            children: [
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'Play sound on',
-                      style: TextStyle(
-                        fontFamily: 'Inter',
-                        fontSize: 16,
-                        fontWeight: FontWeight.w600,
-                        color: c.ink,
-                      ),
-                    ),
-                    Text(
-                      'Where voice and media come out',
-                      style: TextStyle(fontFamily: 'Inter', fontSize: 12, color: c.ink3),
-                    ),
-                  ],
-                ),
-              ),
-              GhostButton(icon: Icons.close, onPressed: () => Navigator.pop(context)),
-            ],
-          ),
-          const SizedBox(height: 14),
-          FreqCard(
-            clipBehavior: Clip.antiAlias,
-            child: Column(
-              children: [
-                for (int i = 0; i < AudioOutput.values.length; i++)
-                  _OutputRow(
-                    output: AudioOutput.values[i],
-                    selected: AudioOutput.values[i] == current,
-                    first: i == 0,
-                    btName: btName,
-                  ),
-              ],
-            ),
-          ),
-          const SizedBox(height: 12),
-          Center(
-            child: Text(
-              "Pair new headphones in your phone's Bluetooth settings.",
-              style: TextStyle(fontFamily: 'Inter', fontSize: 11, color: c.ink3),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _OutputRow extends StatelessWidget {
-  final AudioOutput output;
-  final bool selected;
-  final bool first;
-  final String btName;
-  const _OutputRow({
-    required this.output,
-    required this.selected,
-    required this.first,
-    required this.btName,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final c = FrequencyTheme.of(context).colors;
-    return Material(
-      color: selected ? c.surface2 : c.surface,
-      child: InkWell(
-        onTap: () => Navigator.pop(context, output),
-        child: Container(
-          padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 14),
-          decoration: BoxDecoration(
-            border: Border(
-              top: first ? BorderSide.none : BorderSide(color: c.line),
-            ),
-          ),
-          child: Row(
-            children: [
-              Container(
-                width: 36,
-                height: 36,
-                decoration: BoxDecoration(
-                  color: selected ? c.accentSoft : c.surface2,
-                  borderRadius: BorderRadius.circular(10),
-                ),
-                alignment: Alignment.center,
-                child: Icon(
-                  output.icon,
-                  size: 16,
-                  color: selected ? c.accentInk : c.ink2,
-                ),
-              ),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      output.label,
-                      style: TextStyle(
-                        fontFamily: 'Inter',
-                        fontSize: 14,
-                        fontWeight: FontWeight.w500,
-                        color: c.ink,
-                      ),
-                    ),
-                    Text(
-                      output.subFor(btName),
-                      style: TextStyle(fontFamily: 'Inter', fontSize: 12, color: c.ink3),
-                    ),
-                  ],
-                ),
-              ),
-              if (selected) Icon(Icons.check, size: 16, color: c.accent),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
